Move media to src
