+++ projects/archiver/src/frontTask.ts
@@ -2,7 +2,7 @@ import { Handler } from 'aws-lambda'
 import { unnest } from 'ramda'
 import { attempt, hasFailed } from '../../backend/utils/try'
 import { frontPath, Image } from '../common'
-import { getImagesFromFront } from '../media'
+import { getImagesFromFront } from './media'
 import { getFront } from './downloader'
 import { IssueTaskOutput } from './issueTask'
 import { bucket } from './s3'
